pyproject.toml
[project]
name = "task-tracker-async"
version = "0.1.0"
description = ""
authors = [
    {name = "skip"}
]
readme = "README.md"
requires-python = "3.12.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "python-ulid[pydantic] (>=3.0.0,<4.0.0)",
    "python-ulid (>=3.0.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "cryptography (>=44.0.2,<45.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
